Import wooshBot from friend's repo that is no longer maintained
This is a bot that I programmed years ago with a friend's help, although it is mostly my work - he just fixed a few critical things and offered to maintain it because he had a more reliable computer at the time. It exists purely as a joke and a reference to XKCD comic number 1627. Since my friend no longer maintains it and I'm proud of the work that went into it, I decided to take it back over.